refactor: convert service-object to function
